Mark WPOptions tests incomplete
Really these tests are very brittle and provide next to no value. We
don't really care what WP functions get called, as long as we see the
desired changes in the database. At the moment this test suite does not
boot wordpress which means mocking was the only option here, but I think
we should reimplement these tests as integration tests which boot
worpdress and actually hit the database.